Update el.json Translate to Greek
I've seen that the el.json file was last updated 6 or 7 years ago, so I took the initiative to translate it myself. I tried to translate from the translate website but I quickly figured out that there was no response and my effort would be in vain. That's why I'm proposing a change with the first 405 lines of the last en-US.json translated in Greek. I plan to translate the full file in the near future. If you approve my work I would lke to have access to this file so that the translation occurs even faster. Cheers!